requirements.txt
accelerate
beautifulsoup4
bitsandbytes
blingfire
hf-transfer
huggingface-hub
loguru
lxml
lxml[html_clean]
openai==1.13.3
sentence_transformers
torch
transformers
peft
llama-index
llama-index-embeddings-huggingface
llama-index-llms-huggingface
llama-index-retrievers-bm25
llama-index-vector-stores-milvus
langchain
langchain-community
langchain-openai
langchain-milvus
newspaper3k
pycountry
vllm